refactor: избавился от поля main в package.json библиотеки
