+++ packages/cli/config/fake-entry.ts
@@ -1,0 +1,1 @@
+// I'am a fake entry. Need this for module federation
