+++ packages/cli/config/get-entry-points.ts
