+++ packages/cli/lib/index.ts
@@ -1,4 +1,4 @@
 export { encodePackageName } from './encode-package-name';
-export { getEntryPoints } from './get-entry-points';
+export { getEntryPoints } from '../config/get-entry-points';
 export { getPackageJsonPath } from './get-package-json-path';
 export { readPackageJson } from './read-package-json';
