+++ packages/core/components/microfrontend/__stories__/microfrontend.stories.ts
@@ -1,7 +1,11 @@
-import type { ExampleProps } from '@colibrijs/example';
 import type { Meta, StoryObj } from '@storybook/react';
 
 import { Microfrontend } from '../microfrontend';
+
+type ExampleProps = {
+  title: string;
+  text?: string;
+};
 
 export type MicrofrontendMeta = Meta<typeof Microfrontend<ExampleProps>>;
 export type Story = StoryObj<typeof Microfrontend<ExampleProps>>;
@@ -10,8 +14,7 @@ export default {
   component: Microfrontend,
   title: 'Microfrontend',
   args: {
-    id: '1',
-    componentName: '_colibrijs__example',
+    componentName: 'Example',
     libraryName: '_colibrijs__example',
     src: `${process.env.EXAMPLE_URL}@colibrijs/example`,
     props: {
