+++ packages/core/package.json
@@ -13,7 +13,6 @@
   },
   "devDependencies": {
     "@jest/globals": "^29.7.0",
-    "@colibrijs/example": "^1.0.0",
     "@colibrijs/wait-storybook": "^1.0.0",
     "@storybook/addon-essentials": "^8.0.8",
     "@storybook/addon-interactions": "^8.0.8",
